// src/app/services/models/create-user-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateUserRequest {
  accountNumber: string;
  birthDate: string;
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  phoneNumber: string;
  username: string;
}